+++ src/regdesk/states/loading_search_result_state.js
